Makefile: download latested tested template file (as described in README.md)
Rationale: So far, only released versions of the templates have been
downloaded in `eisvogel` and/or `cleanthesis` targets. However, these
versions may be partially outdated, which can cause translation to fail
with the pandoc templates in this project. So instead the mentioned
targets will download the latest tested version ...

Unfortunately, `git clone --branch` can only address branches or tags,
but not single commits. Thus, a few more commits have to be loaded to
make sure the commit we are looking for is actually loaded ...