Template: switch to single page mode
In the past (before 2020), students had to submit their work in printed, hardback form. A book-like format with double-sided printing was therefore useful.

Nowadays, the work is only submitted digitally - the blank pages at the end of chapters are a nuisance. The alternation of headers and footers is also no longer appropriate.

fixes #79