Template: do not hard wire file names in default files
the user documents, i.e. the thesis document as well as the bibtex
file and the output file should not be hard wired in the default
files - this way we can change this files easily.

this commit moves the bibtex file from the default files back into
the makefile. the thesis document and the output file were already
specified in the makefile.